feature: Toolset Auth settings
