+++ apps/ai-dial-admin/src/components/SourceField/Endpoints/ToolsetEndpoint.tsx
@@ -4,11 +4,14 @@ import classNames from 'classnames';
 import { FC } from 'react';
 
 import { SOURCE_FIELD } from '@/src/components/SourceField/types';
-import { EntitiesI18nKey, EntityPlaceholdersI18nKey } from '@/src/constants/i18n';
+import { EntitiesI18nKey, EntityFieldsI18nKey, EntityPlaceholdersI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
 import { Toolset } from '@/src/models/dial/toolset';
 
+import DropdownField from '@/src/components/Common/Dropdown/DropdownField';
 import EndpointControl from '@/src/components/EntityMainProperties/BaseProperties/Endpoint/Endpoint';
+import { DropdownItemsModel } from '@/src/models/dropdown-item';
+import { ToolsetTransport } from '@/src/types/toolset';
 
 interface Props {
   entity: Toolset;
@@ -19,9 +22,13 @@ interface Props {
 
 const ToolsetEndpoint: FC<Props> = ({ entity, onChange, prefix, isModal }) => {
   const t = useI18n();
+  const transportOptions: DropdownItemsModel[] = [
+    { id: ToolsetTransport.HTTP, name: ToolsetTransport.HTTP.toUpperCase() },
+    { id: ToolsetTransport.SSE, name: ToolsetTransport.SSE.toUpperCase() },
+  ];
 
   return (
-    <div className={classNames(!isModal && 'lg:w-[35%]')}>
+    <div className={classNames('flex flex-col gap-6', !isModal && 'lg:w-[35%]')}>
       {prefix ? (
         <EndpointControl
           id="endpoint"
@@ -47,6 +54,13 @@ const ToolsetEndpoint: FC<Props> = ({ entity, onChange, prefix, isModal }) => {
           onChange={(endpoint) => onChange({ ...entity, endpoint })}
         />
       )}
+      <DropdownField
+        fieldTitle={t(EntityFieldsI18nKey.transport)}
+        selectedValue={entity.transport || ToolsetTransport.SSE}
+        elementId="transport"
+        items={transportOptions}
+        onChange={(transport) => onChange({ ...entity, transport: transport as ToolsetTransport })}
+      />
     </div>
   );
 };
